Updated keymap (#3)
* Updated keymap

* gh module

* rollback gh change

* update upload-artifact

* update upload-artifact

* fix typo

* fix la in nav

* disable lite build; disable macro; use display-name

* disable combos that use macros

* reattempt new sir struct

* move west

* add zephyr/module

* fix build yaml

* attempt to fix west build

by using a shield instead of config

* fix zephyr/module.yml; rename arm, add shield

* Add keymap to board/arm; update build.yaml

Apparently a keymap is needed (likely based on a config arg). Maybe this is why the board cannot be found?

* change board back to zaphod

renaming the board is an extensive change. just leave it as zaphod for now

* Add snippet:studio-rpc-usb-uart and studio_unlock

* disable studio for now

* rm combo labels; rm studio key

* fix deprecated properties

* Fix keymap error; add tapdance

* fix lshift typo

* fix tapdance

hyper macro does not appear to work alone (little rusty) -- needs kp

* fix dumb action typo

* fix thumbs